requirements.txt
Django
django-jazzmin
pillow
weasyprint
django-import-export
django-simple-captcha
djangorestframework